CID 714542: Unintentional sign extension to 64 bits.
Fixed by changing type of val from u_long to u_int32_t. val only gets
passed to htonl which takes a uint32_t, so no need for it to be u_long.

+++ src/kits/network/libbind/inet/inet_addr.c
@@ -107,7 +107,7 @@ inet_addr(const char *cp) {
  */
 int
 inet_aton(const char *cp, struct in_addr *addr) {
-	u_long val;
+	u_int32_t val;
 	int base, n;
 	char c;
 	u_int8_t parts[4];
